+++ utils/src/main/java/io/atomix/utils/memory/Memory.java
@@ -49,11 +49,11 @@ public interface Memory {
     /**
      * Rounds the count to the nearest power of two.
      */
-    public static long toPow2(int size) {
+    public static long toPow2(long size) {
       if ((size & (size - 1)) == 0) {
         return size;
       }
-      int i = 128;
+      long i = 128;
       while (i < size) {
         i *= 2;
         if (i <= 0) {
